+++ src/prelude.rs
@@ -1,5 +1,4 @@
 pub use channel::{mpmc, mpsc, spmc, spsc};
-pub use incin::Incinerator;
 pub use map::Map;
 pub use queue::Queue;
 pub use set::Set;